There is a lot of commonality between different types of model * Added a NewModels::Base class * Made LegacyWrapper more cunning by templating on global vector of models itself
